Split code executions components
